aac frontend CRUD complete

+++ Components/DonneesRefPopUp/DonneesRefPopUp.tsx
@@ -71,7 +71,11 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
 
     const toast = useRef<Toast>(null);
     const deleteToast = useRef<Toast>(null);
+    const updateStatusToast = useRef<Toast>(null);
     const emptyValueErrorToast = useRef<Toast>(null);
+
+    const headlessToast = useRef<Toast>(null);
+
 
     const [visible, setVisible] = useState(true);
     const [editRefData, setEditRefData] = useState(false)
@@ -83,7 +87,6 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
     const [position, setPosition] = useState<'center' | 'top' | 'bottom' | 'left' | 'right' | 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right'>('top');
     // const [position, setPosition] = useState<string>('center');
     const [changesConfirmed, setChangesConfirmed] = useState(false)
-    const [toastState, setToastState] = useState(false)
 
     const [fetchedDonneesRefsData, setFetchedDonneesRefsData] = useState<fetchedDonneesRefsDataStructure[]>([])
 
@@ -91,12 +94,15 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
 
     const [checked, setChecked] = useState<boolean>(false);
 
+    const [isDataBeingDeleted, setIsDataBeingDeleted] = useState(false)
 
-    // fonction ajoutant les nouvelles donnees referentielles
-    const handleAddNewDonneesRef = () => {}
-    
-    // fonction en charge de la recuperation des donnees referentielles existantes
-    const handleGetDonneesRef = () => {}
+    const [confirmationDialogVisibility, setConfirmationDialogVisibility] = useState<boolean>(false)
+    const [confirmationDialogMessage, setConfirmationDialogMessage] = useState<string>('')
+
+    const [updateStatusData, setUpdateStatusData] = useState({
+        newStatus: false,
+        aacCode: 0
+    })
 
     const [products, setProducts] = useState([
         {
@@ -121,10 +127,9 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
     // state for new donnees referentielles form
     const [newDonneesRefFormState, setNewDonneesRefFormState] = useState(false)
 
-    const [isAddNewDataFormVisible, setAddNewDataFormState] = useState(false)
-
     const [newDataValue, seNewDataValue] = useState<string>('')
 
+    // montrer les messages selon si les donnees ont ete modifier ou ajouter
     const footerContent = isEditFormVisible ? (
         <div>
             <Button label="Annulez" icon="pi pi-times" onClick={() => {
@@ -172,45 +177,74 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
     ) 
     ;
 
+    const confirmDialogFooterContent = (
+        <div>
+            <Button label="Annuler" icon="pi pi-times" onClick={() => {
+                setConfirmationDialogVisibility(false); 
+                setIsDataBeingDeleted(false)
+            }} className="p-button-text" />
+            <Button 
+            label="Confirmer" 
+            icon="pi pi-check" 
+            onClick={() => {
+                if(isDataBeingDeleted) {
+                    deleteDonneesReferentiellesData()
+                    setConfirmationDialogVisibility(false)
+                    setIsDataBeingDeleted(false)
+                    return
+                } 
+                if(!isDataBeingDeleted) {
+                    handleAnneeAcademiqueStatusChange()
+                    setConfirmationDialogVisibility(false)
+                    setIsDataBeingDeleted(false)
+                }
+                setIsDataBeingDeleted(false)
+            }} 
+            autoFocus 
+            />
+        </div>
+    );
+
     useEffect(() => {
         setTimeout(() => {
             setChangesConfirmed(false)
         }, 4000)
     }, [changesConfirmed])
-
-    const handleAddNewDonneesReferentielle = () => {
-        if(!newDataValue.trim()) {
-            toast.current && toast.current.show({severity:'error', summary: 'Erreur', detail:`Le champs ne peut etre vide`, life: 3000});
-            return
-        }
-    }
     
     const show = (position: any) => {
         setPosition(position);
         setEditFormState(true);
     };
 
-    const accept = () => {
+    const handleAnneeAcademiqueStatusChange = async () => {
 
-        deleteDonneesReferentiellesData()
+        try {
 
-        deleteToast.current 
-        && 
-        deleteToast.current.show({ severity: 'info', summary: 'Suppresion reussie', detail: 'Données supprimer avec succes', life: 3000 });
-    }   
+            let dataToBeSubmitted = {
+                aacModifierPar: "yannickwnz",
+                aacStatus: updateStatusData.newStatus
+            }
+            
+            const response = await axios.put(`${backendApi}/api/anneeAcademique/${updateStatusData.aacCode}`, dataToBeSubmitted)
+
+            console.log(response.data)
+            if(response.status === 200) {
+                fetchingDonnessReferentielles()
+                if(updateStatusData.newStatus) {
+                    showSuccess('Annee activer avec succes')
+                } else {
+                    showSuccess('Annee desactivee avec succes')
+                }
+            }
+
+        } catch (error) {
+            console.log(error)
+        }
+
+    }
 
 
-    const confirmDelete = () => {
-        confirmDialog({
-            message: 'Êtes vous sûre de vouloir supprimer cette donnée?',
-            header: 'Confirmez la suppression',
-            icon: 'pi pi-info-circle',
-            acceptClassName: 'p-button-danger',
-            accept,
-            // reject
-        });
-    };
-
+    // function qui montre les messages de succes lorsque une donnee a ete mise a jour, creer supprimer avec succes ...
     const showSuccess = (details: string) => {
         toast.current && toast.current.show({severity:'success', summary: 'Succes', detail:`${details}`, life: 3000});
         // toast.current && toast.current.show({severity:'success', summary: 'Succes', detail:'Mis a jour effectue', life: 3000});
@@ -324,6 +358,7 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
             if(response.status === 200) {
                 fetchingDonnessReferentielles()
                 setSelectedDataCode(null)
+                showSuccess('Données supprimer avec succes')
             }
 
         } catch (error) {
@@ -337,12 +372,10 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
         fetchingDonnessReferentielles()
     }, [donnesRef])
 
-
     return (
         <>
-
-        
             {/* dialog containing form input that edit existing donnees referentielles */}
+            {/* dialog contenant l'input qui va servir a mettre a jour la donnee referentielles selectionnee */}
             {isEditFormVisible
                 && 
                 <Dialog 
@@ -371,6 +404,7 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
             }
 
             {/* dialog containing form that added new donnees referentielles */}
+            {/* dialog contenant l'input qui va creer une nouvelle donnee referentielle */}
             {newDonneesRefFormState 
                 && 
                 <Dialog 
@@ -393,6 +427,7 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
             }
 
             {/* dialog containing lists of selected donnees referentielles */}
+            {/* dialog contenant la liste des donnees referentielles */}
             <Dialog 
                 header={`Mise a jour ${donnesRef}`} 
                 visible={visible} 
@@ -424,10 +459,8 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
                             functionSettingEditFormInputValue={setValue}
                         />}
 
-                        
-            
-
                     </div>
+
                     {typeOfDataFetched === typeOfDataFetchedEnums.AnneeAcademiqueType && <div>
                         {fetchedDonneesRefsData.length > 0 ? <div className='table-wrapper'>
                                 <table>
@@ -448,7 +481,23 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
                                                         onIcon="pi pi-check" 
                                                         offIcon="pi pi-times" 
                                                         checked={data.aacStatus ? true : false}
-                                                        onChange={(e:  ToggleButtonChangeEvent) => setChecked(e.value)} 
+                                                        onChange={(e:  ToggleButtonChangeEvent) => {
+                                                            setChecked(e.value)
+                                                            console.log(e.target.value)
+                                                            setUpdateStatusData({
+                                                                newStatus: e.target.value,
+                                                                aacCode: data.aacCode
+                                                            })
+                                                        }} 
+                                                        onClick={() => {
+                                                            setConfirmationDialogVisibility(true)
+                                                            // console.log(data.aacStatus)
+                                                            if(data.aacStatus) {
+                                                                setConfirmationDialogMessage('Etes vous sure de vouloir desactiver cette annee academique ?')
+                                                            } else {
+                                                                setConfirmationDialogMessage("Etes vous sure de vouloir activer cette annee academique? Toutes les autres annees seront automatiquement desactivees")
+                                                            }
+                                                        }}
                                                         className={`w-9rem h-2rem ${data.aacStatus && 'activeStatus'} `}
                                                     />
 
@@ -468,8 +517,11 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
                                                             className="pi pi-trash"
                                                             style={{ fontSize: '1.2rem', color: 'crimson', fontWeight: 'bold' }}
                                                             onClick={() => {
-                                                                confirmDelete()
+                                                                setIsDataBeingDeleted(true)
                                                                 setSelectedDataCode(data.aacCode)
+                                                                // confirmDelete()
+                                                                setConfirmationDialogMessage('Êtes vous sûre de vouloir supprimer cette donnée ?')
+                                                                setConfirmationDialogVisibility(true)
                                                             }}
                                                         >
                                                         </i>
@@ -492,9 +544,27 @@ export const DonneesRefPopUp = ({donnesRef, popUpState, setPopUpState, setDonnee
 
             </Dialog>
 
+            {/* dialog showing confirmation message*/}
+            <Dialog header="Confirmer votre action" visible={confirmationDialogVisibility} style={{ width: '30vw' }} onHide={() => {if (!visible) return; setConfirmationDialogVisibility(false); setIsDataBeingDeleted(false) }} footer={confirmDialogFooterContent}>
+                {confirmationDialogMessage && <div className='flex'>
+                    <span>
+                        <i className="pi pi-info-circle" style={{ fontSize: '2rem', marginRight: '.5em' }}></i>
+                    </span>
+                    <p className="m-0" style={{ fontSize: '1.1rem' }}>
+                        {/* Etes vous sure de vouloir activer cette annee academique?
+                    
+                        Toutes les autres annees seront automatiquement desactivees */}
+                        {confirmationDialogMessage}
+                    </p>
+                </div>}
+
+            </Dialog>
+
             {/* {changesConfirmed && <Toast ref={toast} />} */}
-            { <Toast ref={toast} />}
+            <Toast ref={toast} />
             <Toast ref={deleteToast} />
+            <Toast ref={updateStatusToast} />
+
         </>
     )
